FridayDiscount and BaseDiscount fix

+++ src/main/java/org/example/Uppgift1/Discount.java
@@ -2,7 +2,7 @@ package org.example.Uppgift1;
 
 public interface Discount {
 
-    int apply(Product product);
+    double apply(Product product);
     String getDescription(Product product);
 
 }
